add, edit, fix layout

+++ src/context/ProductContext.tsx
@@ -13,12 +13,23 @@ const ProductContextProvider = ({children} : {children: React.ReactNode}) => {
     }, [])
 
     const onHandleDelete = async (id: number) => {
+     if(confirm('Bạn có muốn xóa không ?')){
       const { data } = await axios.delete(`http://localhost:3000/products/${id}`);
       setProducts(products.filter((item) => item.id !== id))
+     }
+    }
+
+    const onHandleAdd = async (product: IProduct) => {
+      const { data } = await axios.post(`http://localhost:3000/products`, product);
+      setProducts([...products, data]);
+    }
+    const onHandleEdit = async (product: IProduct) => {
+      const { data } = await axios.put(`http://localhost:3000/products/${product.id}`, product);
+      setProducts(products.map((item => item.id === product.id ? product : item )))
     }
   return (
     <div>
-      <ProductContex.Provider value={{products, setProducts, onHandleDelete}}>{children}</ProductContex.Provider>
+      <ProductContex.Provider value={{products, setProducts, onHandleDelete, onHandleAdd, onHandleEdit}}>{children}</ProductContex.Provider>
     </div>
   )
 }
